extract getDataAndCount function to a common hook for consumption in useSearch and useQuickSearch hooks. Add a constant for the election year
