+++ src/constants/common.js
@@ -1,3 +1,4 @@
 export const API_BATCH_SIZE = 10
 export const CANDIDATES = 'candidates'
 export const CONTRIBUTORS = 'contributors'
+export const ELECTION_YEAR = 2020
