+++ src/hooks/useQuickSearch.js
@@ -1,29 +1,17 @@
 import { useCallback, useState } from 'react'
-import { API_BATCH_SIZE } from '../constants'
+
+import { API_BATCH_SIZE, ELECTION_YEAR } from '../constants'
+
+import { useApi } from './useApi'
 
 const constructQuickSearchUrl = ({ url, searchTerm, limit, offset }) =>
-  `/api/${searchTerm}/2020?limit=${limit}&offset=${offset}`
+  `/api/${searchTerm}/${ELECTION_YEAR}?limit=${limit}&offset=${offset}`
 
 export const useQuickSearch = () => {
-  const [hasError, setHasError] = useState(false)
+  const { hasError, setHasError, getDataAndCount } = useApi()
   const [results, setResults] = useState([])
   const [resultsCount, setResultsCount] = useState(0)
   const [resultsOffset, setResultsOffset] = useState(0)
-
-  const getDataAndCount = useCallback(
-    async (url) => {
-      setHasError(false)
-      try {
-        const response = await fetch(url)
-        const { data, count } = await response.json()
-        return { data, count }
-      } catch (e) {
-        console.log(e)
-        setHasError(true)
-      }
-    },
-    [setHasError]
-  )
 
   const fetchQuickSearchData = useCallback(
     async ({ searchTerm, limit = API_BATCH_SIZE, offset = 0 } = {}) => {
@@ -42,7 +30,7 @@ export const useQuickSearch = () => {
         console.log(e)
       }
     },
-    [getDataAndCount]
+    [getDataAndCount, setHasError]
   )
 
   const fetchInitialQuickSearchData = useCallback(
